admin: WarningMessageType refactored to MessageType with "message_level" option
- it is now possible to use also "info" messages in the admin forms
