+++ packages/framework/src/Resources/views/Admin/Form/message.html.twig
@@ -1,0 +1,9 @@
+{% block message_label %}{% endblock message_label %}
+{% block message_widget %}
+    <div class="form-line__line form-line__line--no-space">
+        <div class="in-message in-message--{{ form.vars.message_level }}">
+            <span class="in-message__icon svg svg-{{ form.vars.message_level }}"></span>
+            {{ data|raw }}
+        </div>
+    </div>
+{% endblock message_widget %}
