+++ project-base/app/config/packages/twig.yaml
@@ -24,6 +24,7 @@ twig:
         - '@ShopsysFramework/Admin/Form/imageuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/localizedFullWidth.html.twig'
         - '@ShopsysFramework/Admin/Form/mailWhitelistFormTheme.html.twig'
+        - '@ShopsysFramework/Admin/Form/message.html.twig'
         - '@ShopsysFramework/Admin/Form/multiLocaleBasicFileuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/multiLocaleFileuploadFields.html.twig'
         - '@ShopsysFramework/Admin/Form/numberSlider.html.twig'
@@ -38,7 +39,6 @@ twig:
         - '@ShopsysFramework/Admin/Form/rolesType.html.twig'
         - '@ShopsysFramework/Admin/Form/storeOpeningHoursFormTheme.html.twig'
         - '@ShopsysFramework/Admin/Form/transportPricesFormTheme.html.twig'
-        - '@ShopsysFramework/Admin/Form/warningMessage.html.twig'
         - 'Admin/Form/grapesJsMailType.html.twig'
         - 'Admin/Form/grapesJsType.html.twig'
         - 'Admin/Form/languageConstantFormTheme.html.twig'
